alerts and edit product

+++ src/Components/ProductCard/ProductCard.jsx
@@ -1,5 +1,106 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { AddToCartIcon } from '../Icon';
+// import React, { useState, useEffect, useContext } from 'react';
+// import { AddToCartIcon } from '../Icon';
+// import './Product.css';
+// import { SERVER_URL } from '../../Constants';
+// import { useSelector } from 'react-redux';
+// import { useCart } from '../Shopping/CartContext';
+
+// export const ProductCard = () => {
+//     const { addToCart } = useCart();
+//     const [products, setProducts] = useState([]);
+//     const [loading, setLoading] = useState(true);
+//     const user = useSelector((state) => state?.user?.user);
+
+
+//     const handleRemoveClick = async (product) => {
+//         try {
+//             const response = await fetch(`${SERVER_URL}Producto/EliminarProducto/${product.idProducto}`, {
+//                 method: 'DELETE',
+//                 headers: {
+//                     'Content-Type': 'application/json',
+//                 },
+//                 body: JSON.stringify({ idProducto: product.idProducto }),
+//             });
+//             const data = await response.json();
+//             if (data && data.mensaje === 'ok') {
+//                 console.log('Producto eliminado:', product);
+//             } else {
+//                 console.error('Hubo un error al eliminar el producto:', data);
+//             }
+//         } catch (error) {
+//             console.error('Hubo un error en la solicitud:', error);
+//         }
+//     }
+
+//     const handleClick = (product) => {
+//         addToCart(product);
+//     }
+
+//     useEffect(() => {
+//         const fetchProducts = async () => {
+//             try {
+//                 const response = await fetch(`${SERVER_URL}Producto/ListaProducto`);
+//                 const data = await response.json();
+//                 if (data && data.mensaje === 'ok') {
+//                     setProducts(data.response.reverse());
+//                     console.log('Productos:', data.response);
+//                     setLoading(false);
+//                 } else {
+//                     console.error('Hubo un error al obtener los productos.');
+//                 }
+//             } catch (error) {
+//                 console.error('Hubo un error en la solicitud:', error);
+//             }
+//         };
+
+//         fetchProducts();
+
+//         //Actualizar productos cada 1 segundo
+//         const intervalId = setInterval(fetchProducts, 1000);
+
+//         // Limpiar intervalo cuando el componente se desmonta o se actualiza
+//         return () => clearInterval(intervalId);
+//     }, []);
+
+//     return (
+//         <>
+//             {loading ? (
+//                 <h3>Cargando...</h3>
+//             ) : (
+//                 products.map((productItem) => (
+//                     <figure key={productItem.idProducto}>
+//                         <div className='card_product'>
+//                             <img className='image_product' src={productItem.imagen} alt={productItem.nombre} />
+//                             <p className='name_product'>{productItem.nombre}</p>
+//                             <p className='price_discount'><strong>Precio: </strong><em> $</em> {productItem.precio} 1Kg</p>
+//                             <div className='agregarbotona'>
+//                                 {!user?.idTienda && (
+//                                     <button className='button-addToCartIcon' onClick={() => handleClick(productItem)}>
+//                                         <AddToCartIcon />
+//                                         <p>Agregar</p>
+//                                     </button>
+//                                 )}
+//                                 {/* <button className='button-addToCartIcon' onClick={() => handleClick(productItem)}>
+//                                     <AddToCartIcon />
+//                                     <p>Agregar</p>
+
+//                                 </button> */}
+//                                 <div className='divproduct'>
+//                                 {user?.idTienda && <button className='button-remove' onClick={() => handleRemoveClick(productItem)}>
+//                                     <p className='eliminarproduct'>Eliminar</p>
+//                                 </button>}
+//                                 </div>
+//                             </div>
+//                         </div>
+//                     </figure>
+//                 ))
+//             )}
+//         </>
+//     );
+// }
+
+
+import React, { useState, useEffect } from 'react';
 import './Product.css';
 import { SERVER_URL } from '../../Constants';
 import { useSelector } from 'react-redux';
@@ -9,32 +110,96 @@ export const ProductCard = () => {
     const { addToCart } = useCart();
     const [products, setProducts] = useState([]);
     const [loading, setLoading] = useState(true);
+    const [editingProduct, setEditingProduct] = useState(null); // Estado para controlar el producto en edición
     const user = useSelector((state) => state?.user?.user);
 
-
     const handleRemoveClick = async (product) => {
+        if (window.confirm('¿Estás seguro de que quieres eliminar este producto?')) {
+            try {
+                const response = await fetch(`${SERVER_URL}Producto/EliminarProducto/${product.idProducto}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({ idProducto: product.idProducto }),
+                });
+                const data = await response.json();
+                if (data && data.mensaje === 'ok') {
+                    console.log('Producto eliminado:', product);
+                    // Actualizar la lista de productos después de eliminar
+                    fetchProducts();
+                } else {
+                    console.error('Hubo un error al eliminar el producto:', data);
+                }
+            } catch (error) {
+                console.error('Hubo un error en la solicitud:', error);
+            }
+        }
+    }
+
+    const handleClick = (product) => {
+        if (!user) {
+            alert('Debes iniciar sesión para agregar productos al carrito.');
+        } else {
+            addToCart(product);
+            alert('El producto se ha agregado al carrito.');
+        }
+    }
+
+    const handleEditClick = (product) => {
+        setEditingProduct(product); // Establecer el producto en edición
+    }
+
+    const handleEditCancel = () => {
+        setEditingProduct(null); // Cancelar la edición del producto
+    }
+
+    const handleEditSubmit = async (updatedProduct) => {
+        if (!updatedProduct || !updatedProduct.idProducto) {
+            console.error('El producto a editar es inválido.');
+            return;
+        }
+    
         try {
-            const response = await fetch(`${SERVER_URL}Producto/EliminarProducto/${product.idProducto}`, {
-                method: 'DELETE',
+            const response = await fetch(`${SERVER_URL}Producto/EditarProducto/${updatedProduct.idProducto}`, {
+                method: 'PUT',
                 headers: {
                     'Content-Type': 'application/json',
                 },
-                body: JSON.stringify({ idProducto: product.idProducto }),
+                body: JSON.stringify(updatedProduct),
             });
+            if (!response.ok) {
+                throw new Error('Error al editar el producto');
+            }
             const data = await response.json();
             if (data && data.mensaje === 'ok') {
-                console.log('Producto eliminado:', product);
+                console.log('Producto editado:', updatedProduct);
+                fetchProducts(); // Actualizar la lista de productos
+                setEditingProduct(null); // Cancelar edición después de guardar
             } else {
-                console.error('Hubo un error al eliminar el producto:', data);
+                console.error('Hubo un error al editar el producto:', data);
             }
         } catch (error) {
             console.error('Hubo un error en la solicitud:', error);
         }
     }
 
-    const handleClick = (product) => {
-        addToCart(product);
-    }
+
+    const [editedFields, setEditedFields] = useState({
+        nombre: '',
+        precio: '',
+        imagen: '',
+        cantidad: '',
+    });
+
+
+    const handleEditChange = (field, value) => {
+        setEditedFields({
+            ...editedFields,
+            [field]: value
+        });
+    };
+
 
     useEffect(() => {
         const fetchProducts = async () => {
@@ -55,11 +220,11 @@ export const ProductCard = () => {
 
         fetchProducts();
 
-        //Actualizar productos cada 1 segundo
-        const intervalId = setInterval(fetchProducts, 1000);
+        // //Actualizar productos cada 1 segundo
+        // const intervalId = setInterval(fetchProducts, 1000);
 
-        // Limpiar intervalo cuando el componente se desmonta o se actualiza
-        return () => clearInterval(intervalId);
+        // // Limpiar intervalo cuando el componente se desmonta o se actualiza
+        // return () => clearInterval(intervalId);
     }, []);
 
     return (
@@ -68,7 +233,7 @@ export const ProductCard = () => {
                 <h3>Cargando...</h3>
             ) : (
                 products.map((productItem) => (
-                    <figure key={productItem.idProducto}>
+                    <div key={productItem.idProducto}>
                         <div className='card_product'>
                             <img className='image_product' src={productItem.imagen} alt={productItem.nombre} />
                             <p className='name_product'>{productItem.nombre}</p>
@@ -76,27 +241,45 @@ export const ProductCard = () => {
                             <div className='agregarbotona'>
                                 {!user?.idTienda && (
                                     <button className='button-addToCartIcon' onClick={() => handleClick(productItem)}>
-                                        <AddToCartIcon />
                                         <p>Agregar</p>
                                     </button>
                                 )}
-                                {/* <button className='button-addToCartIcon' onClick={() => handleClick(productItem)}>
-                                    <AddToCartIcon />
-                                    <p>Agregar</p>
-
-                                </button> */}
-                                <div className='divproduct'>
-                                {user?.idTienda && <button className='button-remove' onClick={() => handleRemoveClick(productItem)}>
-                                    <p className='eliminarproduct'>Eliminar</p>
-                                </button>}
-                                </div>
+                                {user?.idTienda && (
+                                    <>
+                                        <button className='button-edit' onClick={() => handleEditClick(productItem)}>
+                                            <p>Editar</p>
+                                        </button>
+                                        <button className='button-remove' onClick={() => handleRemoveClick(productItem)}>
+                                            <p>Eliminar</p>
+                                        </button>
+                                    </>
+                                )}
                             </div>
                         </div>
-                    </figure>
+                        {editingProduct === productItem && (
+                            <div className="popover" onClick={(e) => e.stopPropagation()}>
+                                <h2>Editar Producto</h2>
+                                <form onSubmit={(e) => e.preventDefault()}>
+                                    <label htmlFor="nombre">Nombre:</label>
+                                    <input type="text" id="nombre" value={editedFields.nombre} onChange={(e) => handleEditChange('nombre', e.target.value)} />
+                                    <label htmlFor="precio">Precio:</label>
+                                    <input type="number" id="precio" value={editedFields.precio} onChange={(e) => handleEditChange('precio', e.target.value)} />
+                                    <label htmlFor="imagen">URL de la Imagen:</label>
+                                    <input type="text" id="imagen" value={editedFields.imagen} onChange={(e) => handleEditChange('imagen', e.target.value)} />
+                                    <label htmlFor="cantidad">Cantidad:</label>
+                                    <input type="number" id="cantidad" value={editedFields.cantidad} onChange={(e) => handleEditChange('cantidad', e.target.value)} />
+                                    <button onClick={handleEditCancel}>Cancelar</button>
+                                    <button onClick={() => handleEditSubmit(editedFields)}>Guardar</button>
+                                </form>
+                            </div>
+                        )}
+                    </div>
                 ))
             )}
         </>
     );
 }
+
+
